setup and some progress

+++ SnakeIO/GameModel.cs
@@ -176,7 +176,7 @@ namespace SnakeIO
             if (message.hasCamera)
             {
                 Shared.Components.Positionable position = entity.GetComponent<Shared.Components.Positionable>();
-                entity.Add(new Shared.Components.Camera(new Rectangle(position.pos.X, position.pos.Y, WIDTH, HEIGHT)));
+                entity.Add(new Shared.Components.Camera(new Rectangle((int)position.pos.X, (int)position.pos.Y, WIDTH, HEIGHT)));
             }
 
             if (message.hasKeyboardControllable)
